Update: rewrite cli

+++ src/tasks/BaseTask.ts
@@ -2,6 +2,8 @@ import path from 'path'
 import fsExtra from 'fs-extra'
 import { spawnSync } from './utils/spawnSync'
 import { onGenCommand } from './utils'
+import * as fs from 'fs'
+import chalk from 'chalk'
 
 export interface BaseTaskConfigType {
     root?: string
@@ -16,6 +18,10 @@ export class BaseTask {
     }
     public onCreateCz(): void {
         const czrcPath = path.join(this.root, '.czrc')
+        if (fs.existsSync(czrcPath)) {
+            console.log(chalk.red('The.czrc file already exists.'))
+            process.exit(1)
+        }
         fsExtra.writeJsonSync(czrcPath, {
             path: 'cz-adapter-eslint'
         })
